# packages/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Civic Relay CAD Integration API
  description: |
    REST API para integración con sistemas CAD (Computer Aided Dispatch) tipo 112.

    Permite a Civic Relay enviar alertas de emergencia directamente a centros de coordinación
    y recibir confirmaciones de despacho de unidades.

    **Seguridad**: Implementa rate limiting, CORS, y autenticación API Key (producción).

    **Mejora #2**: Integración 112/CAD - Especificación OpenAPI
  version: 0.1.0
  contact:
    name: Abraham Haddioui
    email: abraham@civic-relay.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.civic-relay.com
    description: Production server (futuro)

tags:
  - name: CAD
    description: Computer Aided Dispatch operations
  - name: Health
    description: Health check endpoints

paths:
  /health:
    get:
      summary: Health check
      description: Verifica que el servidor API está funcionando correctamente
      tags:
        - Health
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: OK
                  timestamp:
                    type: string
                    format: date-time
                  service:
                    type: string
                    example: civic-relay-cad-api
                  version:
                    type: string
                    example: 0.1.0

  /api/cad/dispatch:
    post:
      summary: Create CAD ticket from emergency message
      description: Envía un mensaje de emergencia al sistema CAD para crear un ticket y despachar unidades
      tags:
        - CAD
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageEnvelope'
      responses:
        '201':
          description: Ticket created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  ticket:
                    $ref: '#/components/schemas/CADTicket'
                  message:
                    type: string
                    example: Ticket created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/cad/tickets:
    get:
      summary: Get all CAD tickets
      description: Lista todos los tickets activos en el sistema CAD
      tags:
        - CAD
      responses:
        '200':
          description: List of tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                    example: 42
                  tickets:
                    type: array
                    items:
                      $ref: '#/components/schemas/CADTicket'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/cad/ticket/{id}:
    get:
      summary: Get specific CAD ticket
      description: Obtiene los detalles de un ticket específico por su ID
      tags:
        - CAD
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: TKT-20240909-001
      responses:
        '200':
          description: Ticket details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  ticket:
                    $ref: '#/components/schemas/CADTicket'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/cad/ticket/{id}/status:
    patch:
      summary: Update ticket status
      description: Actualiza el estado de un ticket (para operaciones de dispatcher)
      tags:
        - CAD
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [NEW, DISPATCHED, EN_ROUTE, ON_SCENE, RESOLVED, CLOSED]
                  example: DISPATCHED
                notes:
                  type: string
                  example: "Unidad FIRE-01 despachada"
      responses:
        '200':
          description: Ticket updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  ticket:
                    $ref: '#/components/schemas/CADTicket'
                  message:
                    type: string
                    example: Ticket updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/cad/stats:
    get:
      summary: Get CAD system statistics
      description: Estadísticas agregadas del sistema CAD (KPIs operacionales)
      tags:
        - CAD
      responses:
        '200':
          description: CAD statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  stats:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 42
                      byStatus:
                        type: object
                        additionalProperties:
                          type: integer
                        example:
                          NEW: 5
                          DISPATCHED: 12
                          EN_ROUTE: 8
                          ON_SCENE: 10
                          RESOLVED: 7
                      byPriority:
                        type: object
                        additionalProperties:
                          type: integer
                        example:
                          CRITICAL: 3
                          HIGH: 15
                          MEDIUM: 18
                          LOW: 6
                      avgResponseTime:
                        type: string
                        example: "4.2 min"
                      unitsDispatched:
                        type: integer
                        example: 28

components:
  schemas:
    MessageEnvelope:
      type: object
      required:
        - id
        - payloadType
        - payload
        - priority
        - ttl
        - createdAt
        - sender
      properties:
        id:
          type: string
          example: "msg-uuid-123"
        payloadType:
          type: string
          enum: [EMERGENCY_ALERT, RESOURCE_REQUEST, STATUS_UPDATE, COORDINATION, GENERAL]
          example: EMERGENCY_ALERT
        payload:
          type: object
          description: Payload específico según el tipo de mensaje
          example:
            type: FIRE
            location:
              lat: 40.4168
              lon: -3.7038
            severity: EXTREME
            description: "Incendio forestal activo - Evacuación inmediata"
        priority:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
          example: CRITICAL
        ttl:
          type: integer
          description: Time to live en segundos
          example: 300
        maxHops:
          type: integer
          example: 5
        createdAt:
          type: string
          format: date-time
        sender:
          type: object
          properties:
            deviceId:
              type: string
            userId:
              type: string
            organizationId:
              type: string
        verificationState:
          type: string
          enum: [UNVERIFIED, PENDING, VERIFIED, REJECTED]
          example: VERIFIED

    CADTicket:
      type: object
      required:
        - id
        - incidentType
        - priority
        - status
        - location
        - reportedAt
      properties:
        id:
          type: string
          example: "TKT-20240909-001"
        incidentType:
          type: string
          example: "FIRE"
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
          example: CRITICAL
        status:
          type: string
          enum: [NEW, DISPATCHED, EN_ROUTE, ON_SCENE, RESOLVED, CLOSED]
          example: DISPATCHED
        location:
          type: object
          properties:
            address:
              type: string
              example: "Calle Mayor 123, Madrid"
            coordinates:
              type: object
              properties:
                lat:
                  type: number
                  example: 40.4168
                lon:
                  type: number
                  example: -3.7038
        description:
          type: string
          example: "Incendio forestal activo - Evacuación inmediata"
        reportedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        unitsDispatched:
          type: array
          items:
            type: string
          example: ["FIRE-01", "MEDICAL-03", "POLICE-12"]
        dispatchNotes:
          type: string
          example: "Unidad FIRE-01 despachada a las 14:32"
        originalMessage:
          $ref: '#/components/schemas/MessageEnvelope'

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid message format
              details:
                type: string
                example: Message body and payload are required

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Ticket not found
              ticketId:
                type: string
                example: TKT-123

    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Too many requests, please try again later.

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Failed to create ticket
              details:
                type: string
                example: Database connection failed

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key para autenticación (producción)

security:
  - ApiKeyAuth: []